tests/requirements.txt
beautifulsoup4>=4.12
lxml>=5.0
google-generativeai>=0.8
